+++ CHANGELOG.md
@@ -21,6 +21,7 @@
 ## [3.7.0](https://github.com/BLKSerene/Wordless/releases/tag/3.7.0) - ??/??/2025
 ### 🎉 New Features
 - Settings: Add Settings - Tables - Miscellaneous Settings
+- Utils: Add Wordless's character tokenizer
 - Work Area: Add Work Area - Table Settings - Show total
 
 ### 🔧 Bugfixes